backend/requirements.txt
flask
pymysql
google-cloud-storage
